fixed out of the blue

+++ so_long/libft/ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: nbouhali <nbouhali@student.1337.ma >       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/23 17:51:56 by nbouhali          #+#    #+#             */
-/*   Updated: 2022/10/26 02:00:14 by nbouhali         ###   ########.fr       */
+/*   Updated: 2023/02/11 23:02:59 by nbouhali         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strchr(const char *s, int c)
 {
-	size_t	a;
+	int		a;
 	char	*mbdla;
 
 	a = 0;
