+++ so_long/libft/ft_strnstr.c
@@ -6,16 +6,16 @@
 /*   By: nbouhali <nbouhali@student.1337.ma >       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/23 17:52:22 by nbouhali          #+#    #+#             */
-/*   Updated: 2022/10/26 02:09:04 by nbouhali         ###   ########.fr       */
+/*   Updated: 2023/02/11 23:02:36 by nbouhali         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
+char	*ft_strnstr(const char *haystack, const char *needle, int len)
 {
-	size_t	a;
-	size_t	b;
+	int		a;
+	int		b;
 	char	*tben;
 	char	*lbra;
