+++ so_long/libft/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: nbouhali <nbouhali@student.1337.ma >       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/23 17:52:29 by nbouhali          #+#    #+#             */
-/*   Updated: 2023/01/14 23:21:03 by nbouhali         ###   ########.fr       */
+/*   Updated: 2023/02/11 23:01:55 by nbouhali         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	size_t			b;
+	int				b;
 	char			*subs;
 	unsigned int	test;
 
@@ -22,14 +22,14 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	{
 		return (0);
 	}
-	if (len > ft_strlen(s))
+	if ((int)len > ft_strlen(s))
 		len = ft_strlen(s);
 	b = 0;
 	test = ft_strlen(s);
 	subs = malloc(len * sizeof(char) + 1);
 	if (!subs)
 		return (0);
-	while (b < len && test >= start)
+	while (b < (int)len && test >= start)
 	{
 		subs[b] = s[start];
 		b++;
